Data Object Modeling Phase
