+++ Development Project/Sparcpoint.Core/Models/Product.cs
@@ -2,16 +2,18 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace Sparcpoint.Core.Entities
+namespace Sparcpoint.Core.Models
 {
-    [Table("[Instances].[Categories]")]
-    public class Category
+    [Table("[Instances].[Products]")]
+    public class Product
     {
         [Key]
         public int InstanceId { get; set; }
 
         public string Name { get; set; }
         public string Description { get; set; }
+        public string ProductImageUris { get; set; }
+        public string ValidSkus { get; set; }
         public DateTime CreatedTimestamp { get; set; }
     }
 }
