+++ Development Project/Sparcpoint.Core/Models/ProductAttributes.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations.Schema;
 
-namespace Sparcpoint.Core.Entities
+namespace Sparcpoint.Core.Models
 {
     [Table("[Instances].[ProductAttributes]")]
     public class ProductAttributes
