+++ Development Project/Sparcpoint.Core/Sparcpoint.Core.csproj
@@ -18,10 +18,4 @@
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
   </ItemGroup>
 
-  <ItemGroup>
-    <None Remove="Objects\" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Objects\" />
-  </ItemGroup>
 </Project>
